# features/domain-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/domain": {
            "post": {
                "description": "create new domain on namecheap",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "domain"
                ],
                "summary": "create new domain on namecheap",
                "parameters": [
                    {
                        "type": "string",
                        "description": "domain",
                        "name": "domain",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "chatid",
                        "name": "chatid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/domain/price": {
            "get": {
                "description": "Get the price of the domain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "domain"
                ],
                "summary": "Get the price of the domain",
                "parameters": [
                    {
                        "type": "string",
                        "description": "domain",
                        "name": "domain",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "chatid",
                        "name": "chatid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/nameservers": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve the name server information for a given domain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NameServer"
                ],
                "summary": "Retrieve name server information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Domain name",
                        "name": "domain",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chat ID",
                        "name": "chatid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/main.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ApiResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Change the name server information for a given domain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NameServer"
                ],
                "summary": "Change name server information",
                "parameters": [
                    {
                        "description": "Change name server request",
                        "name": "changeNameServerRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.ChangeNameServerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/main.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ApiResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.ApiResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "object",
                    "properties": {
                        "error": {
                            "type": "string"
                        }
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "main.ChangeNameServerRequest": {
            "type": "object",
            "required": [
                "chatid",
                "domain",
                "nameservers"
            ],
            "properties": {
                "chatid": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "nameservers": {
                    "type": "string"
                }
            }
        }
    }
}